More robustly use effect override internals
The effect override ABI is unstable. The current way it's being
used is somewhat problematic: In release mode, it may silently
stop working or apply the wrong effects and in debug mode it
asserts. Instead, use the julia-level APIs, which have keyword
arg constructors that are slightly more stable and will at least
error consistently.